major refractor + tab restrictions
program was rewritten with the dual-panel system in mind, as the previous implimentation was just a broken 1-panel system cobbled together into a 2-pannel test bed.

The code has been made much more readable and uneeded files have been deleted.

A full backup of the previous version is avalible at /Dev/Proto2.zip

This current version has all of the same behavior as the last version with the following addiitons:
- User is unable to make "entombed tabs", tabs that are in the middle or end of a line
- User is unable to indent a line further than one level above the line before it to prevent undrawable trees