Single-panel prototype
Design has been modified to accomplish the "one panel" design. Program is mostly functional but there are some substantial edge cases that are making the experience less than ideal

TODO: impliment rules on when the user is allowed to hit enter (i.e. never mroe than one new empty line at a time per group)

TODO: fix program so that it recognizes when the carrat has moved via the mouse, as currently the only way to steer the carrat is with arrow keys

TODO:
fix edge case where creating a new line causes the carrat to teleport to the end of the document (could this be an issue of the carrat not updating during the writing of \n and \t)?