single-panel created via stacking + undo splt
in a moment of sheer brilliance, i realized i could just take the two-parallel approach and stack the text areas on top of each other with transparency. this would alow me to create the desired effect with way less work and way less calculation time.

in order for this to work you have to be able to catch the copy event and replace it with new offsets, but because of all the work i did on the Edit-In-Place version i realized i can calculate those offsets quite easily.

Ultimately, I now have a working single-panel version!

All thats left to do now is to block illegal "enter" commands