Fixed indentation loss
Document no longer supports pasting inputs with glyphs of length 8 because doing so was causing indentation to be lost in certain cases.

Basically, "│       " was getting replaced with \t before "│   " and "    ". When a "│   " immediately preceeds a "    " block, the result is indistinguishable from a "│       " block, so the program was correctly turning it into a \t.

TODO: The whole tree really shouldn't be getting calculated every 1000ms. I should change this so that it checks if the raw buffer has actually changed and only THEN does it recalculate the executive buffer (outside of a keypress)