updated auto-update solution + baked main
Previously I was rebuilding the entire tree every 1000ms. Now, whenever the value of the buffer changes for any reason I rebuild the tree.

This means calculations are not done when nothing has changed but does mean that the buffer is calculated 3 times for each keypress. This should be improved.