increased URL tax from 256 to 512
I need to make future-proofing decisions NOW in the timeline. I have increased the size of URL to have 512 characters reserved for things other than the data payload in case I need them. This also makes sense as it leaves 50% of the "standard" 1024 length for data and 50% for everything else.

That being said, given the current 8192 implimentation, more than 95% of the URL is still dedicated to data.