restarted faux bold/italic solutions
a Markdown() object is now attatched to window.main. Calling window.main.test() in console demonstrates that the markdown schema is capable of converting text between four alphabets; normal, bold, italic, and both.

Given this advancement, now all that is left to do is impliment the regex checking to find !.*! and ".*" to mark where to convert and when