smart compression/encoding system implimented
See "URI-mannager.js". File acts as a self-contained mechanism for pushing data to and from the URL in a more robust way. Formerly, it was just /program.html?data=X, but now it is /program.html?enc=X&cmpr=Y&data=Z, such that X is the encoding scheme used, Y is the compression algorithm used, and Z is the data that has been compressed and encoded in those formats.

This is neccessary as I plan to switch the RTN over to LZMA2 compression rather than the current zlib, and need a way to handle backwards compatibility in the future. For notes created with ZLIB, I can deal with the single case by saying "if no enc or cmpr, assume ZLIB", but if I had two depreciated systems I wouldn't be able to tell the difference. Explicitly specifying these allows me to make sure things stay readable in the future.

NOTE: The otag-handler is not equipped to handle anything but zlib compression at present, this may get tricky...