otag overhal  + enable LZMA2
Otag handling for arbitray encoding/compression was starting to become not worth it. As a compomise, I am now compressing just the first line of the document (assumed to be the title) in `?title=`, always in ZLIB. Link previews will no longer contain the note's contents, but will contain that title.

This is a good middle-ground between link preview usability and complexity. Moderate data compression for the relatively small title is sufficent if it means we can use LZMA2 for the actual data payload.

Additionally, the otag-handler.php was revised to only modify the sections that change, other values were internalized into program.html.

LZMA2 compression is also enabled in this commit.