suport decoding via JS puppateer
By migrating to a JS based decomrpession/decoding solution, i can easily support any compresison/encoding schemes I create in the future. This is accomplished via NodeJS Puppeteer, which accesses a dummy website, decomrpessor.php, which after a few secods writes the decompressed data to the sceen to collect for use in metatag processing.

This solution is... slow and not great. I need to look into ways to optimize it, but it works for now.