Reimpliment Otag-handler
So i was being dumb before and really overcomplicating this. Instead of creating a fake webpage to evaluate javascript via terminal.... why dont I just evaluate javascript at terminal?

So that's what im doing now. NodeJS is used to execute "decomrpessor.js" which is a bare-bones version of the URI mannager with PAKO and LZMAJS embedded in the file. This javascript is caleld from the otag-handler.php to rewrite the document.

Now that ims not overcomplicating this, all of the slowness is gone and everything is back to executing almost instantly.