various error handling improvements
significant overhaul to how errors are handled by the RTN. this experiement was conducted by using zzuf (http://caca.zoy.org/wiki/zzuf) to fuzz various links to identify edge behavior in the system. As seen in bench1.txt, before these updates 76% of inputs resulted in an unhandled error. As seen in bench2.txt, now just 0.073% of all inputs result in an unhandled error.

This was accomplished by using try,catch blocks in the decompressor and URI manager such that in the event of a error the contents of the screen are literally replaced with the error. Although this means the "this link was not created by the rtn" message almost never occurs anymore, it means the information that is compunicated is much more useful.

I am still aware that my fuzzing is not comprehensive, if I place a bunch of ampersands all over the place in the URL the system still results in an unhandled error over 50% of the time. I plan to address that issue soon(tm)