impliment actual markdown
ExeBuffer.update does a whole bunch of regular expressions to insert <b>, <i>, <del> into the document. This allows for actual, real markdown to a limited degree. This also simplifies my issues with the fact that not all computers had the esoteric fonts necessary for the faux markdown.

I'm also adding special <span style="color: blue;"> tags for the tree glyhs to keep it cyan, now that the source textarea draws transparent text.