remove html tags from data= and fix otag-handler
any html tags included in the text output are removed from the string that is pushed to the URL. This cuts down on space and ensures weird stuff doesnt happen in link previews.

moreover, the otag handler now properly detects documents that do not have a body (entirely on one line) and sets the <meta description="&nbsp;">. when this string is empty, link previews crash. by explicitly making sure it is *something*, that issue goes away.