fix ul edgecase
consider the line "* goals *". this line should clearly be italiciesed, but the old implimentation would consume the first asterisk for the bullet point, making italics impossible. this version fixes that issue, though does currently mean that LINES WHERE MORE THAN ONE ASTERISK IS PRESENT CAN NEVER BE BULLETED which is an issue and should be fixed