consume misc. instructions from launchpage
instead of blasting users with a bunch of information on the landing page, all this information has been moved to separate pages that can be accessed by special links that are included on the default landing page. this makes it a lot easier to go into further detail about mechanics and syntax without fear of overpopulating the home page.

this is accomplished via the new Schema.redir function which takes in a parameter from a [](#.*) link to navigate to a static documentation sub-page. All subpages are just regular ./program.html pages.