redirect to primary mirror
Added logic to check if the primary copy of the RTN at https://snailien.ddns.net/RTN is online. If the primary copy is online there, all backup copies of the RTN should redirect the user to the same document there. In the event that the primary copy is not availible, users will not be redirected.