make all links open new tab by default
it is preferable to open a new tab than overwrite the current document with a link on accident. no longer will users need to be holding control to open in a new tab, it does it by default.

this change applies to static links, markdown links, and executive (yellow) links